requirements.txt
crawl4ai>=0.4.0
outscraper>=2.2.0
flask>=3.0.0
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
playwright>=1.40.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
rich>=13.7.0
apscheduler>=3.10.4
